fix: Fiz as alterações solicitadas na forma de como os arquivos são organzados e criei o arquivo options.tsx para reutilizar a função e colocar a personalização do tab navigator.
