+++ src/Screens/Perfil/Styles.ts
@@ -1,0 +1,6 @@
+import styled from "styled-components/native";
+
+export const ViewGeneric = styled.View`
+  flex: 1;
+  background-color: #170044;
+`;
