--- src/Styles/StylesApp/StylesApp.ts
@@ -1,5 +1,0 @@
-import styled from 'styled-components/native'
-
-export const ViewApp = styled.View`
-  flex: 1;
-`;
